+++ libsensors/lsm330dlc_gyroscope.c
@@ -118,7 +118,7 @@ int lsm330dlc_gyroscope_activate(struct smdk4x12_sensors_handlers *handlers)
 
 	data = (struct lsm330dlc_gyroscope_data *) handlers->data;
 
-	rc = sysfs_value_write(data->path_enable, 1);
+	rc = sysfs_value_write(data->path_enable, (int64_t) 1);
 	if (rc < 0) {
 		ALOGE("%s: Unable to write sysfs value", __func__);
 		return -1;
@@ -141,7 +141,7 @@ int lsm330dlc_gyroscope_deactivate(struct smdk4x12_sensors_handlers *handlers)
 
 	data = (struct lsm330dlc_gyroscope_data *) handlers->data;
 
-	rc = sysfs_value_write(data->path_enable, 0);
+	rc = sysfs_value_write(data->path_enable, (int64_t) 0);
 	if (rc < 0) {
 		ALOGE("%s: Unable to write sysfs value", __func__);
 		return -1;
